angler: typec: change the current advertisement to 3A
when a non-powered charger is connected, CC is connected to VBUS
through a resistor Rp (in the charger or in the cable)and there is
a capacitor between VBUS and ground(in the charger), the loop
may bring the CC voltage to a range that tusb320 can detect an UFP
attach and turn on VBUS.
The Type-C to A cable with 10k resistor(non-Standard value) and
Type-C charger with 10k resistor (Standard value) will cause the
issue. Changing the CC termination Rp (by changing the current
advertisement to 3A) can make some improvement, fix the issue of
all A chargers and most of Type-C chargers.
Change the current advertisement to 3A and turn it back to default
before enabling VBUS.

BUG: 24193079

Change-Id: I50578aeb6a65a6ec511ef9d677202a57b50d3569
Signed-off-by: Wei Li <sirius.liwei@huawei.com>
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>