angler: charging: Fix the bug of not power off when charger is present
If the battery soc is 0%, the device should shutdown itself to protect
peripheral chip. But with charger plug in, the protection does not work.
This may cause peripheral chips damaged in low frequency. So we need
to add shutdown action while the battery voltage is low and charger is
present.

BUG=25169707

Signed-off-by: zhen yuan <richard.yuan@huawei.com>

Change-Id: Icb1afece0f86fdb26e37b5a472303a7e69d309ed
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>